requirements.txt
alembic==1.7.5
anyio==3.2.1
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
bleach==3.3.0
blinker==1.4
CacheControl==0.12.10
cachetools==5.0.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.4
django-mathfilters==1.0.0
docker==5.0.3
dominate==2.6.0
entrypoints==0.3
filelock==3.6.0
firebase-admin==5.2.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gcloud==0.18.3
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.0
googleapis-common-protos==1.54.0
graphviz==0.17
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.2
idna==2.10
importlib-metadata==3.10.1
importlib-resources==5.4.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jws==0.1.3
kiwisolver==1.3.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.3
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
npm==0.1.1
numpy==1.20.3
oauth2client==4.1.3
optional-django==0.1.0
packaging==21.0
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.20.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.12.0
Pygments==2.9.0
pymongo==4.0.1
pyparsing==2.4.7
Pyrebase4==4.5.0
pyrsistent==0.18.0
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
PyYAML==6.0
pyzmq==22.1.0
razorpay==0.2.0
requests==2.27.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.8
scipy==1.6.3
scrapinghub==2.3.1
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.4
virtualenv==20.13.1
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
WTForms==3.0.1
zipp==3.5.0
